moving ASSIST switch to utils
