+++ openpype/hosts/nuke/api/__init__.py
@@ -1,4 +1,3 @@
-import os
 from .workio import (
     file_extensions,
     has_unsaved_changes,
@@ -52,8 +51,6 @@ from .utils import (
     get_colorspace_list
 )
 
-ASSIST = bool(os.getenv("NUKEASSIST"))
-
 __all__ = (
     "file_extensions",
     "has_unsaved_changes",
@@ -98,7 +95,5 @@ __all__ = (
     "create_write_node",
 
     "colorspace_exists_on_node",
-    "get_colorspace_list",
-
-    "ASSIST"
+    "get_colorspace_list"
 )
